requirements.txt
# Athena Trading Signal Detection System
# Main Requirements

# Data processing
polars>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0

# API clients
requests>=2.28.0
websockets>=11.0.0
aiohttp>=3.8.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Technical analysis
ta-lib>=0.4.25

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Monitoring and logging
loguru>=0.7.0

# Alerts
twilio>=8.0.0

# Web framework (for dashboard)
streamlit>=1.28.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0

# Date/time handling
pendulum>=2.1.0
pytz>=2023.3

# Configuration
pyyaml>=6.0

# HTTP client enhancements
httpx>=0.24.0
